# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API AlexTa",
        "description": "Documentation de l'API de l'application AlexTa",
        "contact": {
            "email": "support@alexta.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/supplements": {
            "get": {
                "tags": [
                    "SupplementGabarit"
                ],
                "summary": "Afficher tous les suppléments gabarits",
                "operationId": "b0075909b7ad106cd5a8539fc0eed592",
                "responses": {
                    "200": {
                        "description": "Liste des suppléments gabarits",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SupplementGabarit"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur"
                    }
                }
            },
            "post": {
                "tags": [
                    "SupplementGabarit"
                ],
                "summary": "Créer un supplément gabarit",
                "operationId": "626a203e5095fa24a4f70b7180b4e3dc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "montant"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "Type B"
                                    },
                                    "montant": {
                                        "type": "integer",
                                        "example": 150
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Supplément gabarit créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SupplementGabarit"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/api/supplements/{id}": {
            "get": {
                "tags": [
                    "SupplementGabarit"
                ],
                "summary": "Afficher un supplément gabarit spécifique",
                "operationId": "5866e6b6b9ab5d33af04af70acce02bd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du supplément gabarit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplément gabarit trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SupplementGabarit"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Supplément gabarit non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "SupplementGabarit"
                ],
                "summary": "Mettre à jour un supplément gabarit",
                "operationId": "4f2d179902b86feff8e051088209f80e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du supplément gabarit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "Type C"
                                    },
                                    "montant": {
                                        "type": "integer",
                                        "example": 200
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Supplément gabarit mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SupplementGabarit"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Supplément gabarit non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "SupplementGabarit"
                ],
                "summary": "Supprimer un supplément gabarit",
                "operationId": "d85f3ed541c657c37d7a046cd46de706",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du supplément gabarit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplément gabarit supprimé avec succès"
                    },
                    "404": {
                        "description": "Supplément gabarit non trouvé"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SupplementGabarit": {
                "title": "Supplement Gabarit",
                "required": [
                    "type",
                    "montant"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "example": "Type A"
                    },
                    "montant": {
                        "type": "integer",
                        "example": 100
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}